backend/requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# PostgreSQL (used in production on Render/Railway)
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Production server
gunicorn==21.2.0

# Environment variables
python-dotenv==1.0.1
